+++ doc/development/release_notes.rst
@@ -16,6 +16,9 @@ Keep up-to-date with the latest changes to the Arcade library by the release not
 
   * Placeholder to put something about sections here.
 
+* Add parallax example: :ref:`parallax`
+* Add :func:`arcade.get_angle_degrees` function.
+
 2.6.10
 ------